xml BUGFIX type size limits

+++ src/xml.c
@@ -372,7 +372,7 @@ lyxml_parse_qname(struct lyxml_ctx *xmlctx, const char **prefix, uint32_t *prefi
         LY_CHECK_RET(lyxml_parse_identifier(xmlctx, &start, &end));
     }
 
-    if (((uint32_t)(end - start)) > UINT32_MAX) {
+    if (((uint64_t)(end - start)) > UINT32_MAX) {
         LOGVAL(xmlctx->ctx, NULL, LYVE_SYNTAX, "XML qualified name too long.");
         return LY_EINVAL;
     }
